requirements.txt
streamlit==1.45.1
fpdf==1.7.2
google-generativeai==0.8.5
Markdown==3.8
nltk==3.9.1
torch==2.7.0+cpu
torchaudio==2.7.0+cpu
torchvision==0.22.0+cpu
weasyprint==65.1
